LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/radio-double-activate-pseudo-expected.txt



FAIL <input type=radio> shouldn't have the :active pseudo element after pressing the spacebar twice. assert_equals: If the radio doesn't have the :active pseudo selector, nothing else should either. expected null but got Element node <input type="radio" id="radioinput"></input>
